+++ source/refined-github.ts
@@ -213,7 +213,7 @@ import './features/hide-markdown-diff';
 import './features/rgh-sponsor-button';
 import './features/rgh-linkify-features';
 import './features/conversation-activity-filter';
-import './features/quick-review-comment-deletion';
+import './features/delete-review-comments-faster';
 
 // Add global for easier debugging
 (window as any).select = select;
